+++ bin/fireshell.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 var fireshell = require('../dist/index');
-fireshell.main();
+fireshell.fireshellCLI();