+++ src/index.js
@@ -1,9 +1,40 @@
 const inquirer = require('inquirer');
 const fs = require('fs');
+const { Command } = require('commander')
 const path = require('path');
+
 const { processQuery } = require('./query');
+const { processOutput, createConfigFile } = require('./utils');
+const { firerun } = require('./firerun');
+const {
+    setConfigCallback,
+    resetCallback
+} = require('./callbacks');
+
+function fireshellCLI() {
+    const program = new Command();
+    program
+        .command('set-config <path> <db>')
+        .option('-u, --url type', 'Url for firebase realtime database')
+        .action(setConfigCallback);
+    
+    program
+        .command('reset')
+        .action(resetCallback);
+    
+    program
+        .command('firerun <inputFilePath> <outputDir>')
+        .action(firerun)
+    
+    program.parse(process.argv);
+
+    if(process.argv.length <= 2) {
+        main()
+    }
+}
 
 function main(){
+    console.log(__dirname);
     const dirs = fs.readdirSync(__dirname + "/../")
     if(dirs.indexOf("config.json") === -1) {
         inquirer.prompt([
@@ -34,22 +65,14 @@ function main(){
                 console.error(Error('You need to provide the database url in order to connect to the realtime database'));
                 process.exit();
             }
-    
-            fs.writeFile(
-                path.join(__dirname, '../', 'config.json'), 
-                JSON.stringify(
-                    {
-                        db: responses.db,
-                        path: responses.path,
-                        url: responses.url
-                    }
-                ),
-                (err) => {
-                    if(err) {
-                        console.log(err);
-                    }
+            createConfigFile(
+                path.join(__dirname, '../', 'config.json'),
+                {
+                    db: responses.db,
+                    path: responses.path,
+                    url: responses.url
                 }
-            );
+            )
             inputQueries();
         });
     } else {
@@ -68,23 +91,19 @@ function inputQueries() {
         if(output instanceof Error){
             console.error(`\n${output}\n`);
             inputQueries();
-        } else{
-            output.then(data => {
-                if(typeof data === 'string'){
-                    console.log(`\n${data}\n`);
-                    console.log(`\nQuery execution failed\n`);
-                } else if(typeof data === 'object' && data !== null){
-                    console.log(JSON.stringify(data, null, 4));
-                    console.log('\nQuery execution successful\n');
-                } else if(data instanceof Error){
-                    console.error(data);
+        } else {
+            processOutput(output).then(
+                result => {
+                    console.log(JSON.stringify(result, null, 2))
+                    inputQueries();
                 }
-                inputQueries();
-            });
+            );
         }
     });
 }
 
 module.exports = {
-    main
+    main,
+    fireshellCLI,
+    inputQueries
 }