+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes will be documented in this file
 
+## Changelog v1.2.0
+
+### Release Highlights
+
+- Added commands for setting and resetting config file.
+- Added command for running multiple queries at once by taking input from a file.
+
+
 ## Changelog v1.1.0
 
 ### Added